refactor repository cache to support different types
 - restructure
