+++ lib/config/types.ts
@@ -11,7 +11,7 @@ export type RenovateConfigStage =
   | 'pr';
 
 export type RepositoryCacheConfig = 'disabled' | 'enabled' | 'reset';
-export type RepositoryCacheType = 'local' | string;
+export type RepositoryCacheType = 'local' | 'redis' | 's3';
 export type DryRunConfig = 'extract' | 'lookup' | 'full';
 export type RequiredConfig = 'required' | 'optional' | 'ignored';
