+++ lib/util/cache/repository/impl/client/local/local.spec.ts
@@ -1,14 +1,15 @@
 import { promisify } from 'util';
 import zlib from 'zlib';
 import hasha from 'hasha';
-import { fs } from '../../../../../test/util';
-import { GlobalConfig } from '../../../../config/global';
-import { CACHE_REVISION } from '../common';
-import type { RepoCacheData, RepoCacheRecord } from '../types';
-import { CacheClientFactory } from './cache-client-factory';
-import { RepositoryCacheHandler } from './repository-cache-handler';
+import { fs } from '../../../../../../../test/util';
+import { GlobalConfig } from '../../../../../../config/global';
+import { logger } from '../../../../../../logger';
+import { CACHE_REVISION } from '../../../common';
+import type { RepoCacheData, RepoCacheRecord } from '../../../types';
+import { RepositoryCacheImpl } from '../../repository-cache-impl';
+import { CacheClientFactory } from '../cache-client-factory';
 
-jest.mock('../../../fs');
+jest.mock('../../../../../fs');
 
 const compress = promisify(zlib.brotliCompress);
 
@@ -24,13 +25,13 @@ async function createCacheRecord(
   return { revision, repository, payload, hash };
 }
 
-describe('util/cache/repository/impl/local', () => {
+describe('util/cache/repository/impl/client/local/local', () => {
   beforeEach(() => {
     GlobalConfig.set({ cacheDir: '/tmp/cache', platform: 'github' });
   });
 
   it('returns empty object before any data load', () => {
-    const localRepoCache = new RepositoryCacheHandler('some/repo', 'local');
+    const localRepoCache = new RepositoryCacheImpl('some/repo', 'local');
     expect(localRepoCache.getData()).toBeEmpty();
   });
 
@@ -38,7 +39,7 @@ describe('util/cache/repository/impl/local', () => {
     const data: RepoCacheData = { semanticCommits: 'enabled' };
     const cacheRecord = await createCacheRecord(data);
     fs.readCacheFile.mockResolvedValue(JSON.stringify(cacheRecord));
-    const localRepoCache = new RepositoryCacheHandler('some/repo', 'local');
+    const localRepoCache = new RepositoryCacheImpl('some/repo', 'local');
 
     await localRepoCache.load();
 
@@ -53,7 +54,7 @@ describe('util/cache/repository/impl/local', () => {
         semanticCommits: 'enabled',
       })
     );
-    const localRepoCache = new RepositoryCacheHandler('some/repo', 'local');
+    const localRepoCache = new RepositoryCacheImpl('some/repo', 'local');
 
     await localRepoCache.load();
     await localRepoCache.save();
@@ -73,7 +74,7 @@ describe('util/cache/repository/impl/local', () => {
         data: { semanticCommits: 'enabled' },
       })
     );
-    const localRepoCache = new RepositoryCacheHandler('some/repo', 'local');
+    const localRepoCache = new RepositoryCacheImpl('some/repo', 'local');
 
     await localRepoCache.load();
     await localRepoCache.save();
@@ -94,7 +95,7 @@ describe('util/cache/repository/impl/local', () => {
       })
     );
 
-    const localRepoCache = new RepositoryCacheHandler('some/repo', 'local');
+    const localRepoCache = new RepositoryCacheImpl('some/repo', 'local');
     await localRepoCache.load();
 
     expect(localRepoCache.getData()).toBeEmpty();
@@ -102,7 +103,7 @@ describe('util/cache/repository/impl/local', () => {
 
   it('handles invalid data', async () => {
     fs.readCacheFile.mockResolvedValue(JSON.stringify({ foo: 'bar' }));
-    const localRepoCache = new RepositoryCacheHandler('some/repo', 'local');
+    const localRepoCache = new RepositoryCacheImpl('some/repo', 'local');
 
     await localRepoCache.load();
 
@@ -111,7 +112,7 @@ describe('util/cache/repository/impl/local', () => {
 
   it('handles file read error', async () => {
     fs.readCacheFile.mockRejectedValue(new Error('unknown error'));
-    const localRepoCache = new RepositoryCacheHandler('some/repo', 'local');
+    const localRepoCache = new RepositoryCacheImpl('some/repo', 'local');
 
     await localRepoCache.load();
 
@@ -123,7 +124,7 @@ describe('util/cache/repository/impl/local', () => {
     const cacheRecord = createCacheRecord({ semanticCommits: 'enabled' });
     fs.readCacheFile.mockResolvedValueOnce(JSON.stringify(cacheRecord));
 
-    const localRepoCache = new RepositoryCacheHandler('some/repo', 'local');
+    const localRepoCache = new RepositoryCacheImpl('some/repo', 'local');
     await localRepoCache.load();
 
     expect(localRepoCache.getData()).toEqual({});
@@ -132,14 +133,7 @@ describe('util/cache/repository/impl/local', () => {
   it('saves modified cache data to file', async () => {
     const oldCacheRecord = createCacheRecord({ semanticCommits: 'enabled' });
     fs.readCacheFile.mockResolvedValueOnce(JSON.stringify(oldCacheRecord));
-    const localRepoCache = new RepositoryCacheHandler('some/repo');
-
-    // TODO: remove use implemented, this test file should test local client only
-    CacheClientFactory.reset();
-    new RepositoryCacheHandler('some/repo', 'redis://'); // coverage
-    CacheClientFactory.reset();
-    new RepositoryCacheHandler('some/repo', 'https://s3'); // coverage
-
+    const localRepoCache = new RepositoryCacheImpl('some/repo');
     await localRepoCache.load();
     const data = localRepoCache.getData();
     data.semanticCommits = 'disabled';
@@ -151,6 +145,12 @@ describe('util/cache/repository/impl/local', () => {
     expect(fs.outputCacheFile).toHaveBeenCalledWith(
       '/tmp/cache/renovate/repository/github/some/repo.json',
       JSON.stringify(newCacheRecord)
+    );
+    // TODO: remove once implemented, this test file should test local client only
+    CacheClientFactory.reset();
+    new RepositoryCacheImpl('some/repo', 'redis'); // coverage
+    expect(logger.info).toHaveBeenCalledWith(
+      `Repository cache type: redis not supported using type "local" instead`
     );
   });
 });
