+++ lib/util/cache/repository/init.ts
@@ -1,6 +1,6 @@
 import { GlobalConfig } from '../../../config/global';
 import type { RenovateConfig } from '../../../config/types';
-import { RepositoryCacheHandler } from './impl/repository-cache-handler';
+import { RepositoryCacheImpl } from './impl/repository-cache-impl';
 import { resetCache, setCache } from '.';
 
 /**
@@ -17,14 +17,14 @@ export async function initRepoCache(config: RenovateConfig): Promise<void> {
   }
 
   if (repositoryCache === 'enabled') {
-    const cache = new RepositoryCacheHandler(repository, type);
+    const cache = new RepositoryCacheImpl(repository, type);
     await cache.load();
     setCache(cache);
     return;
   }
 
   if (repositoryCache === 'reset') {
-    const cache = new RepositoryCacheHandler(repository, type);
+    const cache = new RepositoryCacheImpl(repository, type);
     await cache.save();
     setCache(cache);
     return;
